fix: patch gho balanceOf (#183)
vGHO has a special mechanic to handle the discount through stkAAVE, applied on any user action and when fetching the balance via `balanceOf`.
When calculating the users debt balance, the protocol currently uses the scaledBalanceOf * index, to save roundtrips to token addresses. While for "usual" tokens, this optimisation makes sense, it will overaccount the vGHO balance by ignoring the discount since the last interaction, as scaledBalanceOf doesn't take it into account.
Therefore, this patch relies on the fact that `!virtualAccounting == GHO` to enter a code-branch opting out of the optimization.

This approach was chosen as it does not require an update on vGHO itself. We consider a more permanent solution in an upcoming upgrade.

---

In other code paths like liquidations, the code already opts out of the optimization as it relies on balanceOf.