fix: improve presision of HF calculation
Certora managed to prove an occasion where the supply of 1 wei could reduce a users hf to to precision loss on avgLt calculation. While this is an extreme adge case on very small amounts, it makes sense to just not lose the precision if possible.

Moving the hf check on borrow at the end, reduces code duplication and makes the operation more equivalent to e.g. withdraw.