fix: remove wrong lt check
When liquidating a collateral, the liquidationThreshold of the collateral is no longer validated.
As long as an asset is flagged as collateralEnabled on a user position it can be liquidated.
This already was the implicit behavior on previous versions of Aave as for an asset to be enabled as collateral, it must have a liquidationThreshold != 0 for the current user configuration.
Removing the redundant validation, allows to maintain that same behavior on Aave v3.6, as on v3.6 it is possible for an asset to be collateral inside an eMode, while outside the liquidationThreshold is 0.

For a more extensive overview over the collateral enablement rules & the rules for switching an eMode, please check the v3.6-properties.